third_party/blink/web_tests/platform/generic/tables/hittesting/filltable-levels-expected.txt

SUCCESS!!!
Hit cell at row: 0, column: 0: SUCCESS
Hit cell at row: 0, column: 1: SUCCESS
Hit cell at row: 0, column: 2: SUCCESS
Hit cell at row: 1, column: 0: SUCCESS
Hit cell at row: 1, column: 1: SUCCESS
Hit cell at row: 1, column: 2: SUCCESS
Hit cell at row: 2, column: 0: SUCCESS
Hit cell at row: 2, column: 1: SUCCESS
Hit cell at row: 2, column: 2: SUCCESS
